Addressed Steve's comments about renaming and adjusting tests
