+++ hadoop-tools/hadoop-aws/src/main/java/org/apache/hadoop/fs/s3a/impl/streams/InputStreamType.java
@@ -44,7 +44,7 @@ public enum InputStreamType {
    * The analytics input stream.
    */
   Analytics("analytics", c ->
-      new S3ASeekableInputStreamFactory());
+      new AnalyticsStreamFactory());
 
   /**
    * Name.
