+++ hadoop-tools/hadoop-aws/src/test/java/org/apache/hadoop/fs/s3a/statistics/ITestS3AFileSystemStatistic.java
@@ -44,7 +44,7 @@ public class ITestS3AFileSystemStatistic extends AbstractS3ATestBase {
    */
   @Test
   public void testBytesReadWithStream() throws IOException {
-    // Assertions will fail as {@link S3ASeekableInputStream}
+    // Assertions will fail as {@link AnalyticsStream}
     // do not support S3AFileSystemStatistics yet.
     skipIfAnalyticsAcceleratorEnabled(getConfiguration(),
         "S3SeekableStream does not support File System Statistics");
